Persist post history for a given day (#20)
* use mutex instead of synchronized lock

* extract persistince logic to own service

* persist postHistory for given day
