+++ active-member-discord-bot/src/main/kotlin/org/regenagcoop/model/ActiveMemberConfig.kt
@@ -4,16 +4,31 @@ import org.regenagcoop.discord.model.ChannelId
 import org.regenagcoop.discord.model.GuildId
 import org.regenagcoop.discord.model.RoleId
 import org.regenagcoop.discord.model.UserId
+import java.time.LocalDate
 
 data class ActiveMemberConfig(
     val guildId: GuildId, // aka "guild id"
     val excludedUserIds: Set<UserId>, // bots
     val roleConfigs: List<RoleConfig>,
     val downgradeMessageConfig: DowngradeMessageConfig,
+    val persistenceConfig: PersistenceConfig,
 ) {
-    val maxWindowSize: Int = roleConfigs.flatMap {
+    private val maxWindowSize: Int = roleConfigs.flatMap {
             listOf(it.keepRoleConfig.windowSize, it.addRoleConfig.windowSize)
     }.max()
+
+    /**
+     * The earliest date we need to consider when scanning for post history.
+     * Since we can't scan back for reaction history, this is only used for post history
+     */
+    fun computeEarliestScanDate(today: LocalDate): LocalDate {
+        val daysToLookBack = maxWindowSize
+        return today.minusDays(daysToLookBack - 1L)
+    }
+
+    data class PersistenceConfig(
+        val channel: ChannelId,
+    )
 
     data class RoleConfig(
         val roleId: RoleId,
